% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeticDys_both.R
\name{EdgeticDys_both}
\alias{EdgeticDys_both}
\title{Get dysregulated edges in cancer}
\usage{
#Identifying the dysregulated edges in both two conditions
Input.exp=GetExampleData(exampleData="Exp.input")
#obtain the sample label
Cancer_s=GetExampleData(exampleData="Cancer_s")
Normal_s=GetExampleData(exampleData="Normal_s")
#obtain the protein interaction network
Network=GetExampleData(exampleData="network")
#identify the dysregulated edges
DysCN=EdgeticDys_CN(Input.exp,Network,thr=0.01)
Input.exp=Input.exp[,Cancer_s]
DysC=EdgeticDys_CN(Input.exp,Network,thr=0.01)
Dys.net=EdgeticDys_both(DysCN,DysC)
}
\arguments{
\item{DysCN}{The dysregulated network in cancer vs normal}

\item{DysC}{The dysregulated network for specific cancer sample}
}
\description{
Identify the dysregulated edges for each sample
}
