% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeticDriver.R
\name{getEdgeticDriver}
\alias{getEdgeticDriver}
\title{Identify the edgetic driver mutations}
\usage{
#Identifying the mutation mediated-dysregulated edges in both two conditions
#obtain the data for gene expression.
Input.exp=GetExampleData(exampleData="Exp.input")
#obtain the sample label
Cancer_s=GetExampleData(exampleData="Cancer_s")
Normal_s=GetExampleData(exampleData="Normal_s")
#obtain the protein interaction network
Network=GetExampleData(exampleData="network")
#obtain the mutation data
Mut=GetExampleData(exampleData="mut")
DysCN=EdgeticDys_CN(Input.exp,Network,thr=0.01)
Input.exp2=Input.exp[,Cancer_s]
DysC=EdgeticDys_CN(Input.exp2,Network,thr=0.01)
Dys.net=EdgeticDys_both(DysCN,DysC)
Driver.mut=getEdgeticDriver(Dys.net,Mut,alpha=0.05,n.sim=1000)
}
\arguments{
\item{Dysnet}{Dysregulated network for samples}

\item{Mut}{The mutation files for samples}

\item{alpha}{The significance level}

\item{n.sim}{The simulation times defined by users}
}
\value{
MutPPI the mutation perturbed PPIs in cancer samples
}
\description{
Identify the driver mutations that perturbe the protein interaction network
}
