% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExampleData.R
\name{GetExampleData}
\alias{GetExampleData}
\title{Load example datasets}
\usage{
#obtain the data for gene expression.
Input.exp=GetExampleData(exampleData="Exp.input")
#obtain the sample label
Cancer_s=GetExampleData(exampleData="Cancer_s")
Normal_s=GetExampleData(exampleData="Normal_s")
#obtain the protein interaction network
Network=GetExampleData(exampleData="network")
#obtain the mutation data
Mut=GetExampleData(exampleData="mut")
}
\arguments{
\item{exampleData}{Indicate which type of default datasets to import}
}
\value{
return the default datasets required by the users
}
\description{
The default example files
}
