% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grubbs.flag.R
\name{grubbs.flag}
\alias{grubbs.flag}
\title{Identify the outliers}
\usage{
grubbs.flag(x, thr)
}
\arguments{
\item{x}{Vector}

\item{thr}{significant level}
}
\description{
Identify the outliers
}
